+++ stock_operating_unit/__init__.py
@@ -1,6 +1,4 @@
-# -*- coding: utf-8 -*-
-# © 2016-2017 Eficent Business and IT Consulting Services S.L.
-# © 2016-2017 Serpent Consulting Services Pvt. Ltd.
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
+
 from . import model
 from .hooks import update_operating_unit_location
