+++ stock_operating_unit/i18n/stock_operating_unit.pot
@@ -1,0 +1,123 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* stock_operating_unit
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: stock_operating_unit
+#: code:addons/stock_operating_unit/model/stock_warehouse.py:50
+#, python-format
+msgid "Configuration Error. The Operating Unit of the Warehouse and the Location must be the same. "
+msgstr ""
+
+#. module: stock_operating_unit
+#: code:addons/stock_operating_unit/model/stock_location.py:51
+#, python-format
+msgid "Configuration error. The Company in the Stock Location and in the Operating Unit must be the same."
+msgstr ""
+
+#. module: stock_operating_unit
+#: code:addons/stock_operating_unit/model/stock_picking.py:32
+#, python-format
+msgid "Configuration error. The Company in the Stock Picking and in the Operating Unit must be the same."
+msgstr ""
+
+#. module: stock_operating_unit
+#: code:addons/stock_operating_unit/model/stock_warehouse.py:34
+#, python-format
+msgid "Configuration error. The Company in the Stock Warehouse and in the Operating Unit must be the same."
+msgstr ""
+
+#. module: stock_operating_unit
+#: code:addons/stock_operating_unit/model/stock_picking.py:44
+#, python-format
+msgid "Configuration error. The Operating Unit of the picking must be the same as that of the warehouse of the Picking Type."
+msgstr ""
+
+#. module: stock_operating_unit
+#: code:addons/stock_operating_unit/model/stock_location.py:65
+#, python-format
+msgid "Configuration error. The Parent Stock Location must belong to the same Operating Unit."
+msgstr ""
+
+#. module: stock_operating_unit
+#: code:addons/stock_operating_unit/model/stock_move.py:39
+#, python-format
+msgid "Configuration error. The Stock moves must be related to a location (source or destination) that belongs to the requesting Operating Unit."
+msgstr ""
+
+#. module: stock_operating_unit
+#: code:addons/stock_operating_unit/model/stock_location.py:35
+#: code:addons/stock_operating_unit/model/stock_location.py:40
+#, python-format
+msgid "Configuration error. The operating unit should be assigned to internal locations only."
+msgstr ""
+
+#. module: stock_operating_unit
+#: code:addons/stock_operating_unit/model/stock_location.py:24
+#, python-format
+msgid "Configuration error. This location is assigned to a warehouse that belongs to a different operating unit."
+msgstr ""
+
+#. module: stock_operating_unit
+#: model:ir.model.fields,field_description:stock_operating_unit.field_stock_move__operating_unit_dest_id
+msgid "Dest. Location Operating Unit"
+msgstr ""
+
+#. module: stock_operating_unit
+#: model:ir.model,name:stock_operating_unit.model_stock_location
+msgid "Inventory Locations"
+msgstr ""
+
+#. module: stock_operating_unit
+#: model:ir.model,name:stock_operating_unit.model_stock_warehouse_orderpoint
+msgid "Minimum Inventory Rule"
+msgstr ""
+
+#. module: stock_operating_unit
+#: model:ir.model.fields,field_description:stock_operating_unit.field_stock_location__operating_unit_id
+#: model:ir.model.fields,field_description:stock_operating_unit.field_stock_rule__operating_unit_id
+#: model:ir.model.fields,field_description:stock_operating_unit.field_stock_warehouse__operating_unit_id
+#: model_terms:ir.ui.view,arch_db:stock_operating_unit.view_picking_internal_search
+msgid "Operating Unit"
+msgstr ""
+
+#. module: stock_operating_unit
+#: model:ir.model.fields,field_description:stock_operating_unit.field_stock_picking__operating_unit_id
+msgid "Requesting Operating Unit"
+msgstr ""
+
+#. module: stock_operating_unit
+#: model:ir.model.fields,field_description:stock_operating_unit.field_stock_move__operating_unit_id
+msgid "Source Location Operating Unit"
+msgstr ""
+
+#. module: stock_operating_unit
+#: model:ir.model,name:stock_operating_unit.model_stock_move
+msgid "Stock Move"
+msgstr ""
+
+#. module: stock_operating_unit
+#: model:ir.model,name:stock_operating_unit.model_stock_rule
+msgid "Stock Rule"
+msgstr ""
+
+#. module: stock_operating_unit
+#: model:ir.model,name:stock_operating_unit.model_stock_picking
+msgid "Transfer"
+msgstr ""
+
+#. module: stock_operating_unit
+#: model:ir.model,name:stock_operating_unit.model_stock_warehouse
+msgid "Warehouse"
+msgstr ""
+
