+++ stock_operating_unit/model/__init__.py
@@ -1,5 +1,7 @@
-# -*- coding: utf-8 -*-
-# © 2016-2017 Eficent Business and IT Consulting Services S.L.
-# © 2016-2017 Serpent Consulting Services Pvt. Ltd.
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
-from . import stock
+
+from . import stock_location
+from . import stock_move
+from . import stock_picking
+from . import stock_rule
+from . import stock_warehouse
