+++ stock_operating_unit/readme/CONFIGURE.rst
@@ -1,0 +1,4 @@
+To configure this module, you need to:
+
+* Assign Operating Unit to Warehouses.
+* Assign Operating Unit to Stock Locations.
