+++ stock_operating_unit/readme/CONTRIBUTORS.rst
@@ -1,0 +1,3 @@
+* Jordi Ballester Alomar <jordi.ballester@eficent.com>
+* Aaron Henriquez <aheficent@eficent.com>
+* Sudhir Arya <sudhir.arya@serpentcs.com>
