+++ stock_operating_unit/readme/ROADMAP.rst
@@ -1,0 +1,3 @@
+The Manager can see the stock rules of other Operating Units but he can not
+edit them. If he tries to access to one of these stock rules, he will receive
+a configuration error.
