+++ stock_operating_unit/security/stock_security.xml
@@ -1,11 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2016-2017 Eficent Business and IT Consulting Services S.L.
-     Copyright 2016-2017 Serpent Consulting Services Pvt. Ltd.
+<!-- Copyright 2019 Eficent Business and IT Consulting Services S.L.
+     Copyright 2019 Serpent Consulting Services Pvt. Ltd.
      License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl-3.0) -->
-<odoo noupdate="1">
+<odoo>
     <record id="ir_rule_stock_warehouse_allowed_operating_units" model="ir.rule">
         <field name="model_id" ref="stock.model_stock_warehouse"/>
-        <field name="domain_force">['|',('operating_unit_id','=',False),('operating_unit_id','in',[g.id for g in user.operating_unit_ids])]</field>
+        <field name="domain_force">['|',
+            ('operating_unit_id', '=', False),
+            ('operating_unit_id', 'in', user.operating_unit_ids.ids)]
+        </field>
         <field name="name">Warehouses from allowed operating units</field>
         <field name="global" eval="True"/>
         <field eval="0" name="perm_unlink"/>
@@ -16,7 +19,10 @@
 
     <record id="ir_rule_stock_picking_type_allowed_operating_units" model="ir.rule">
         <field name="model_id" ref="stock.model_stock_picking_type"/>
-        <field name="domain_force">['|',('warehouse_id.operating_unit_id','=',False),('warehouse_id.operating_unit_id','in',[g.id for g in user.operating_unit_ids])]</field>
+        <field name="domain_force">['|',
+            ('warehouse_id.operating_unit_id','=', False),
+            ('warehouse_id.operating_unit_id','in',user.operating_unit_ids.ids)]
+        </field>
         <field name="name">Stock Picking Type from allowed operating units</field>
         <field name="global" eval="True"/>
         <field eval="0" name="perm_unlink"/>
@@ -27,8 +33,10 @@
 
     <record id="ir_rule_stock_location_allowed_operating_units" model="ir.rule">
         <field name="model_id" ref="stock.model_stock_location"/>
-        <field name="domain_force">['|', ('operating_unit_id','in',[g.id for g
-            in user.operating_unit_ids]), ('operating_unit_id','=', False)]</field>
+        <field name="domain_force">['|',
+            ('operating_unit_id', 'in',user.operating_unit_ids.ids),
+            ('operating_unit_id','=',False)]
+        </field>
         <field name="name">Stock locations from allowed operating units</field>
         <field name="global" eval="True"/>
         <field eval="0" name="perm_unlink"/>
@@ -39,8 +47,13 @@
 
     <record id="ir_rule_stock_move_allowed_operating_units" model="ir.rule">
         <field name="model_id" ref="stock.model_stock_move"/>
-        <field name="domain_force">['|',('location_id.operating_unit_id','=',False),('location_id.operating_unit_id','in',[g.id for g in user.operating_unit_ids]),
-            '|',('location_dest_id.operating_unit_id','=',False),('location_dest_id.operating_unit_id','in',[g.id for g in user.operating_unit_ids])]</field>
+        <field name="domain_force">['|',
+            ('location_id.operating_unit_id','=',False),
+            ('location_id.operating_unit_id','in',user.operating_unit_ids.ids),
+            '|',
+            ('location_dest_id.operating_unit_id','=',False),
+            ('location_dest_id.operating_unit_id','in',user.operating_unit_ids.ids)]
+        </field>
         <field name="name">Stock moves from allowed operating units</field>
         <field name="global" eval="True"/>
         <field eval="0" name="perm_unlink"/>
@@ -51,7 +64,10 @@
 
     <record id="ir_rule_stock_picking_allowed_operating_units" model="ir.rule">
         <field name="model_id" ref="stock.model_stock_picking"/>
-        <field name="domain_force">['|',('operating_unit_id','=',False),('operating_unit_id','in',[g.id for g in user.operating_unit_ids])]</field>
+        <field name="domain_force">['|',
+            ('operating_unit_id','=',False),
+            ('operating_unit_id','in',user.operating_unit_ids.ids)]
+        </field>
         <field name="name">Stock pickings from allowed operating units</field>
         <field name="global" eval="True"/>
         <field eval="0" name="perm_unlink"/>
@@ -60,4 +76,17 @@
         <field eval="0" name="perm_create"/>
     </record>
 
+    <record id="ir_rule_stock_picking_allowed_picking_type_operating_units" model="ir.rule">
+        <field name="model_id" ref="stock.model_stock_picking"/>
+        <field name="domain_force">['|',
+            ('picking_type_id.warehouse_id.operating_unit_id','=',False),
+            ('picking_type_id.warehouse_id.operating_unit_id','in',user.operating_unit_ids.ids)]
+        </field>
+        <field name="name">Stock pickings from allowed picking types</field>
+        <field name="global" eval="True"/>
+        <field eval="0" name="perm_unlink"/>
+        <field eval="0" name="perm_write"/>
+        <field eval="1" name="perm_read"/>
+        <field eval="0" name="perm_create"/>
+    </record>
 </odoo>
