+++ stock_operating_unit/tests/test_stock_picking.py
@@ -1,6 +1,5 @@
-# -*- coding: utf-8 -*-
-# © 2016-2017 Eficent Business and IT Consulting Services S.L.
-# © 2016-2017 Serpent Consulting Services Pvt. Ltd.
+# © 2019 Eficent Business and IT Consulting Services S.L.
+# © 2019 Serpent Consulting Services Pvt. Ltd.
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 from . import test_stock_operating_unit as test_stock_ou
